requirements.txt
flask>=3.0.0
google-generativeai>=0.8.0
python-pptx>=1.0.0
markdown>=3.7
python-dotenv>=1.0.0
